refactor: 숫자 생성 인터페이스 생성
