+++ src/main/java/sehee/util/numbermaker/NumberMaker.java
@@ -1,0 +1,9 @@
+package sehee.util.numbermaker;
+
+public interface NumberMaker {
+
+    int make();
+
+    int[] makeAllUnique(int length);
+
+}
