+++ src/main/java/sehee/util/numbermaker/RandomNumberMaker.java
@@ -1,35 +1,35 @@
-package sehee.answer;
+package sehee.util.numbermaker;
 
 import java.util.Random;
 
 // package-private
-public class NumberMaker {
+public class RandomNumberMaker implements NumberMaker {
 
     private final Random random;
 
-    public NumberMaker(Random random) {
+    public RandomNumberMaker(Random random) {
         this.random = random;
     }
 
-    int makeRandomOne() {
+    public int make() {
         return random.nextInt(9) + 1;
     }
 
-    int[] makeAllUniqueRandom(int length) {
+    public int[] makeAllUnique(int length) {
         boolean[] isExist = new boolean[10];
         isExist[0] = true;
 
         int[] randomNumbers = new int[length];
         int count = 0;
         while (count < randomNumbers.length) {
-            count = makeAndCountUniqueRandom(isExist, randomNumbers, count);
+            count = makeAndCountUnique(isExist, randomNumbers, count);
         }
 
         return randomNumbers;
     }
 
-    int makeAndCountUniqueRandom(boolean[] isExist, int[] randomNumbers, int count) {
-        int randomNumber = makeRandomOne(); // 1~9
+    private int makeAndCountUnique(boolean[] isExist, int[] randomNumbers, int count) {
+        int randomNumber = this.make(); // 1~9
         if (!isExist[randomNumber]) { // 모두 다른 숫자
             isExist[randomNumber] = true;
             randomNumbers[count++] = randomNumber;
